document-ia-api/pyproject.toml
[tool.poetry]
name = "document-ia-api"
version = "0.1.0"
description = "API minimaliste FastAPI avec authentification API_KEY"
authors = ["Romaric HALDENWANG <romaric.haldenwang@beta.gouv.fr>"]
readme = "README.md"
packages = [
    { include = "api", from = "src" },
    { include = "application", from = "src" },
    { include = "core", from = "src" },
    { include = "infra", from = "src" },
    { include = "schemas", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
fastcrud = "^0.16.0"
sqlalchemy = "^2.0.43"
uvicorn = "^0.24.0"
gunicorn = "^22.0.0"
boto3 = "^1.34.0"
python-magic = "^0.4.27"
python-multipart = "^0.0.6"
asyncpg = "^0.30.0"
greenlet = "^3.0.0"
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
document-ia-infra = {path = "../document-ia-infra" }

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pytest = "^8.4.1"
pytest-asyncio = "^0.23.0"
pre-commit = "^3.6.0"
httpx = "^0.25.0"
pyright = "^1.1.405"
boto3-stubs = {extras = ["essential"], version = "^1.40.31"}


[tool.poetry.scripts]
start = "uvicorn main:app --app-dir document-ia-api --host 0.0.0.0 --port 8000"
dev = "uvicorn main:app --app-dir document-ia-api --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
